+++ cmd/build_options.bat
@@ -1,5 +1,5 @@
 set TARGET=longtail
-set SRC=%BASE_DIR%src\*.c %BASE_DIR%cmd\main.cpp %BASE_DIR%lib\*.c %BASE_DIR%lib\bikeshed\*.c %BASE_DIR%lib\blake2\*.c %BASE_DIR%lib\filestorage\*.c %BASE_DIR%lib\lizard\*.c %BASE_DIR%lib\memstorage\*.c %BASE_DIR%lib\meowhash\*.c %BASE_DIR%lib\xxhash\*.c
+set SRC=%BASE_DIR%src\*.c %BASE_DIR%src\ext\*.c %BASE_DIR%cmd\main.cpp %BASE_DIR%lib\*.c %BASE_DIR%lib\bikeshed\*.c %BASE_DIR%lib\blake2\*.c %BASE_DIR%lib\filestorage\*.c %BASE_DIR%lib\lizard\*.c %BASE_DIR%lib\memstorage\*.c %BASE_DIR%lib\meowhash\*.c %BASE_DIR%lib\xxhash\*.c
 set TEST_SRC=
 set THIRDPARTY_SRC=%THIRDPARTY_DIR%dummy.c
 rem %THIRDPARTY_DIR%lizard\lib\*.c %THIRDPARTY_DIR%lizard\lib\entropy\*.c %THIRDPARTY_DIR%lizard\lib\xxhash\*.c
