+++ src/ext/stb_ds.c
@@ -1,0 +1,2 @@
+#define STB_DS_IMPLEMENTATION
+#include "stb_ds.h"
