+++ test/build_options.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 export TARGET=test
-export SRC="$BASE_DIR/src/*.c $BASE_DIR/lib/*.c $BASE_DIR/lib/bikeshed/*.c $BASE_DIR/lib/blake2/*.c $BASE_DIR/lib/filestorage/*.c $BASE_DIR/lib/lizard/*.c $BASE_DIR/lib/memstorage/*.c $BASE_DIR/lib/meowhash/*.c $BASE_DIR/lib/xxhash/*.c"
+export SRC="$BASE_DIR/src/*.c $BASE_DIR/src/ext/*.c $BASE_DIR/lib/*.c $BASE_DIR/lib/bikeshed/*.c $BASE_DIR/lib/blake2/*.c $BASE_DIR/lib/filestorage/*.c $BASE_DIR/lib/lizard/*.c $BASE_DIR/lib/memstorage/*.c $BASE_DIR/lib/meowhash/*.c $BASE_DIR/lib/xxhash/*.c"
 export TEST_SRC="$BASE_DIR/test/impl_bind.c $BASE_DIR/test/main.cpp $BASE_DIR/test/test.cpp"
 export THIRDPARTY_SRC="$THIRDPARTY_DIR/dummy.c"
 #"$THIRDPARTY_DIR/lizard/lib/*.c $THIRDPARTY_DIR/lizard/lib/entropy/*.c $THIRDPARTY_DIR/lizard/lib/xxhash/*.c $THIRDPARTY_DIR/blake2/sse/blake2s.c"
